requirements.txt
Flask==2.3.2
spotipy==2.23.0
google-generativeai==0.8.3
requests==2.31.0
python-dotenv==1.0.0
